fixing

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: hyeonkki <hyeonkki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/05 16:19:13 by hyeonkki          #+#    #+#             */
-/*   Updated: 2021/05/05 19:42:23 by hyeonkki         ###   ########.fr       */
+/*   Updated: 2021/05/06 22:13:26 by hyeonkki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 void	*ft_memset(void *ptr, int value, size_t num)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	str = ptr;
-	while (i < (int)num)
+	while (i < num)
 		str[i++] = value;
 	return (ptr);
 }
